change main components
