+++ src/stylesheet/public.scss
@@ -599,6 +599,7 @@ input[type='number'] {
             .bottom {
               position: relative;
               z-index: 3;
+              object-fit: initial !important;
             }
           }
           .token-low {
